Add tests to app
