+++ .eslintrc.yml
@@ -2,7 +2,10 @@ env:
   node: true
   commonjs: true
   es2021: true
+  jest: true
 extends: standard
 parserOptions:
   ecmaVersion: latest
 rules: {}
+globals:
+  __ENV: readonly
