+++ __mocks__/env.js
@@ -1,0 +1,16 @@
+module.exports = {
+  DATABASE_HOST: 'localhost',
+  DATABASE_USERNAME: 'username',
+  DATABASE_PASSWORD: 'password',
+  DATABASE_NAME: 'unleash',
+  ADMIN_ACCESS_TOKEN: 'admin-access-token',
+  CLIENT_ACCESS_TOKEN: 'client-access-token',
+  SESSION_SECRET: 'session-secret',
+  KC_HOST: 'http://localhost:8080/auth',
+  KC_REALM: 'realm',
+  KC_CLIENT_ID: 'kc-client-id',
+  KC_CLIENT_SECRET: 'kc-client-secret',
+  KC_ADMIN_ROLES: ['admin'],
+  KC_EDITOR_ROLES: ['editor'],
+  KC_VIEWER_ROLES: ['viewer']
+}
