+++ package.json
@@ -4,10 +4,18 @@
   "main": "index.js",
   "scripts": {
     "start": "NODE_ENV=production node index.js",
-    "test:ci": "echo 'no tests for unleash-docker';"
+    "lint": "eslint .",
+    "lint:fix": "eslint --fix .",
+    "test": "jest"
   },
   "engines": {
     "node": ">=18"
+  },
+  "jest": {
+    "testPathIgnorePatterns": [
+      "/node_modules/",
+      "/integration_test/"
+    ]
   },
   "author": "",
   "license": "ISC",
@@ -23,6 +31,8 @@
     "eslint-config-standard": "^17.1.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-n": "^16.0.1",
-    "eslint-plugin-promise": "^6.1.1"
+    "eslint-plugin-promise": "^6.1.1",
+    "jest": "^29.7.0",
+    "supertest": "^6.3.3"
   }
 }
